Stop discriminating body-accepting methods in client DSL
